requirements.txt
apache-airflow==2.7.1
google-cloud-bigquery==3.12.0
google-cloud-storage==2.13.0
pandas==2.0.3
requests==2.31.0
pyarrow==13.0.0
pandas-gbq==0.19.2
db-dtypes==1.1.1
croniter==1.3.8
python-dateutil==2.8.2